Made Index.cs a partial class with smaller dependents files
